+++ src/routes/fleuve/changelog.md
@@ -2,6 +2,8 @@
 latest: 0.1.2
 ---
 
+# Changelog
+
 ### Version 0.1.2
 
 - Activate new page after creation
@@ -18,8 +20,27 @@ latest: 0.1.2
 - Keep siblings visible when activating a page.
 - Cleaned up store helpers.
 
+# Todo
+
+- Multiple document support
+- Some storage mechanism besides localstorage
+
 <style>
-	h3 {
+	h1 {
+		font-size: 20px;
+		padding: 8px 12px;
+		border-top: 1px solid #ddd;
+		border-bottom: 1px solid #ddd;
+		margin-bottom: 8px;
+		margin-top: 24px;
+	}
+
+	h1:first-child {
+		border-top: none;
+		margin-top: 0px;
+	}
+
+ 	h3 {
 		padding: 0px 12px;
 		margin: 0;
 	}